backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdforbit-backend"
version = "0.1.0"
description = "Production-grade FastAPI backend scaffold for PdfORBIT."
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.14,<2.0",
  "argon2-cffi>=23.1,<26.0",
  "bcrypt>=4.2,<5.0",
  "beautifulsoup4>=4.13,<5.0",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "openpyxl>=3.1,<4.0",
  "orjson>=3.10,<4.0",
  "pdf2image>=1.17,<2.0",
  "pikepdf>=9.7,<10.0",
  "Pillow>=11.0,<13.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyMuPDF>=1.26,<2.0",
  "PyJWT>=2.10,<3.0",
  "pytesseract>=0.3.13,<1.0",
  "python-docx>=1.1,<2.0",
  "python-multipart>=0.0.20,<0.1.0",
  "python-pptx>=1.0,<2.0",
  "pydantic-settings>=2.6,<3.0",
  "redis>=5.2,<6.0",
  "rq>=1.16,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
  "fakeredis[lua]>=2.31,<3.0",
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["build*", "tests*", "var*", "pdforbit_backend.egg-info*"]

[tool.pytest.ini_options]
addopts = "-q -ra"
testpaths = ["tests"]
